requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
sqlalchemy==2.0.36
pydantic-settings==2.7.1
python-dotenv==1.0.1
PyJWT==2.10.1
sse-starlette==2.2.1
